Refactor CategoriesPage component and improve error handling
