+++ client/admin-web/src/api/services/categoryApiService.js
@@ -1,4 +1,4 @@
-import apiClient from "./apiClient.js";
+import apiClient from "../apiClient.js";
 
 export const getCategories = async ({ pageNumber, pageSize }) => {
   const { data } = await apiClient.get("/categories", {
