+++ client/admin-web/src/utilities/dateUtilities.js
@@ -1,0 +1,9 @@
+export function formatDate(dateStr) {
+  if (!dateStr) return "-";
+
+  return new Date(dateStr).toLocaleDateString(undefined, {
+    day: "2-digit",
+    month: "short",
+    year: "numeric",
+  });
+}
